Add unit tests for the container improvements code.

+++ tests/cc/test_c_api.cc
@@ -13,10 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include <fcntl.h>
 #include <dlfcn.h>
 #include <stdint.h>
 #include <string.h>
 #include <sys/mman.h>
+#include <sys/mount.h>
+#include <sys/stat.h>
+#include <sys/types.h>
 #include <unistd.h>
 
 #include "bcc_elf.h"
@@ -29,6 +33,8 @@
 #include "catch.hpp"
 
 using namespace std;
+
+static pid_t spawn_child(void *, bool, bool, int (*)(void *));
 
 TEST_CASE("shared object resolution", "[c_api]") {
   char *libm = bcc_procutils_which_so("m", 0);
@@ -103,6 +109,83 @@ extern "C" int _a_test_function(const char *a_string) {
   return i;
 }
 
+static int setup_tmp_mnts(void) {
+  // Disconnect this mount namespace from its parent
+  if (mount(NULL, "/", NULL, MS_REC|MS_PRIVATE, NULL) < 0) {
+    fprintf(stderr, "unable to mark / PRIVATE: %s\n", strerror(errno));
+    return -1;
+  }
+  // create a new tmpfs mounted on /tmp
+  if (mount("tmpfs", "/tmp", "tmpfs", 0, NULL) < 0) {
+    fprintf(stderr, "unable to mount /tmp in mntns: %s\n", strerror(errno));
+    return -1;
+  }
+
+  return 0;
+}
+
+static int mntns_func(void *arg) {
+  int in_fd, out_fd;
+  char buf[4096];
+  char libpath[1024];
+  ssize_t rb;
+  void *dlhdl;
+
+  if (setup_tmp_mnts() < 0) {
+    return -1;
+  }
+
+  // Find libz.so.1, if it's installed
+  dlhdl = dlopen("libz.so.1", RTLD_LAZY);
+  if (dlhdl == NULL) {
+    fprintf(stderr, "Unable to dlopen libz.so.1: %s\n", dlerror());
+    return -1;
+  }
+
+  if (dlinfo(dlhdl, RTLD_DI_ORIGIN, &libpath) < 0) {
+    fprintf(stderr, "Unable to find origin of libz.so.1: %s\n", dlerror());
+    return -1;
+  }
+
+  dlclose(dlhdl);
+  dlhdl = NULL;
+
+  // Copy a shared library from shared mntns to private /tmp
+  snprintf(buf, 4096, "%s/libz.so.1", libpath);
+  in_fd = open(buf, O_RDONLY);
+  if (in_fd < 0) {
+    fprintf(stderr, "Unable to open %s: %s\n", buf, strerror(errno));
+    return -1;
+  }
+
+  out_fd = open("/tmp/libz.so.1", O_RDWR|O_CREAT|O_EXCL,
+      S_IRWXU|S_IRGRP|S_IXGRP|S_IROTH|S_IXOTH);
+  if (out_fd < 0) {
+    fprintf(stderr, "Unable to open /tmp/libz.so.1: %s\n", strerror(errno));
+    return -1;
+  }
+  memset(buf, 0, sizeof (buf));
+  while ((rb = read(in_fd, buf, sizeof (buf))) > 0) {
+    if (write(out_fd, buf, rb) < 0) {
+      fprintf(stderr, "Write error: %s\n", strerror(errno));
+      return -1;
+    }
+  }
+  close(in_fd);
+  close(out_fd);
+
+  dlhdl = dlopen("/tmp/libz.so.1", RTLD_NOW);
+  if (dlhdl == NULL) {
+    fprintf(stderr, "dlopen error: %s\n", dlerror());
+    return -1;
+  }
+
+  sleep(5);
+  dlclose(dlhdl);
+
+  return 0;
+}
+
 TEST_CASE("resolve symbol addresses for a given PID", "[c_api]") {
   struct bcc_symbol sym;
   void *resolver = bcc_symcache_new(getpid());
@@ -142,6 +225,21 @@ TEST_CASE("resolve symbol addresses for a given PID", "[c_api]") {
     REQUIRE(string(sym.module).find("libc") != string::npos);
     REQUIRE(string("strtok") == sym.name);
   }
+
+  SECTION("resolve in separate mount namespace") {
+    pid_t child;
+    uint64_t addr = 0;
+
+    child = spawn_child(0, true, true, mntns_func);
+    REQUIRE(child > 0);
+
+    void *resolver = bcc_symcache_new(child);
+    REQUIRE(resolver);
+
+    REQUIRE(bcc_symcache_resolve_name(resolver, "/tmp/libz.so.1", "zlibVersion",
+        &addr) == 0);
+    REQUIRE(addr != 0);
+  }
 }
 
 #define STACK_SIZE (1024 * 1024)
@@ -151,10 +249,7 @@ static string perf_map_path(pid_t pid) {
   return tfm::format("/tmp/perf-%d.map", pid);
 }
 
-static int child_func(void *arg) {
-  unsigned long long map_addr = (unsigned long long)arg;
-
-  string path = perf_map_path(getpid());
+static int make_perf_map_file(string &path, unsigned long long map_addr) {
   FILE *file = fopen(path.c_str(), "w");
   if (file == NULL) {
     return -1;
@@ -163,19 +258,56 @@ static int child_func(void *arg) {
   fprintf(file, "%llx 10 right_next_door_fn\n", map_addr + 0x10);
   fclose(file);
 
+  return 0;
+}
+
+static int perf_map_func(void *arg) {
+  string path = perf_map_path(getpid());
+  if (make_perf_map_file(path, (unsigned long long)arg) < 0)
+    return -1;
+
   sleep(5);
 
   unlink(path.c_str());
   return 0;
 }
 
-static pid_t spawn_child(void *map_addr, bool own_pidns) {
+static int perf_map_func_mntns(void *arg) {
+  string path = perf_map_path(getpid());
+
+  if (setup_tmp_mnts() < 0) {
+    return -1;
+  }
+
+  if (make_perf_map_file(path, (unsigned long long)arg) < 0)
+    return -1;
+
+  sleep(5);
+
+  unlink(path.c_str());
+  return 0;
+}
+
+static int perf_map_func_noop(void *arg) {
+  if (setup_tmp_mnts() < 0) {
+    return -1;
+  }
+
+  sleep(5);
+
+  return 0;
+}
+
+static pid_t spawn_child(void *map_addr, bool own_pidns, bool own_mntns,
+    int (*child_func)(void *)) {
   int flags = 0;
   if (own_pidns)
     flags |= CLONE_NEWPID;
+  if (own_mntns)
+    flags |= CLONE_NEWNS;
 
-  pid_t child = clone(child_func, /* stack grows down */ child_stack + STACK_SIZE,
-      flags, (void*)map_addr);
+  pid_t child = clone(child_func,
+      /* stack grows down */ child_stack + STACK_SIZE, flags, (void*)map_addr);
   if (child < 0)
     return -1;
 
@@ -193,7 +325,7 @@ TEST_CASE("resolve symbols using /tmp/perf-pid.map", "[c_api]") {
   pid_t child = -1;
 
   SECTION("same namespace") {
-    child = spawn_child(map_addr, /* own_pidns */ false);
+    child = spawn_child(map_addr, /* own_pidns */ false, false, perf_map_func);
     REQUIRE(child > 0);
 
     void *resolver = bcc_symcache_new(child);
@@ -213,7 +345,24 @@ TEST_CASE("resolve symbols using /tmp/perf-pid.map", "[c_api]") {
   }
 
   SECTION("separate namespace") {
-    child = spawn_child(map_addr, /* own_pidns */ true);
+    child = spawn_child(map_addr, /* own_pidns */ true, false, perf_map_func);
+    REQUIRE(child > 0);
+
+    void *resolver = bcc_symcache_new(child);
+    REQUIRE(resolver);
+
+    REQUIRE(bcc_symcache_resolve(resolver, (unsigned long long)map_addr,
+        &sym) == 0);
+    REQUIRE(sym.module);
+    // child is PID 1 in its namespace
+    REQUIRE(string(sym.module) == perf_map_path(1));
+    REQUIRE(string("dummy_fn") == sym.name);
+    unlink("/tmp/perf-1.map");
+  }
+
+  SECTION("separate pid and mount namespace") {
+    child = spawn_child(map_addr, /* own_pidns */ true, true,
+        perf_map_func_mntns);
     REQUIRE(child > 0);
 
     void *resolver = bcc_symcache_new(child);
@@ -226,6 +375,29 @@ TEST_CASE("resolve symbols using /tmp/perf-pid.map", "[c_api]") {
     REQUIRE(string(sym.module) == perf_map_path(1));
     REQUIRE(string("dummy_fn") == sym.name);
   }
+
+  SECTION("separate pid and mount namespace, perf-map in host") {
+    child = spawn_child(map_addr, /* own_pidns */ true, true,
+        perf_map_func_noop);
+    REQUIRE(child > 0);
+
+    string path = perf_map_path(child);
+    REQUIRE(make_perf_map_file(path, (unsigned long long)map_addr) == 0);
+
+    void *resolver = bcc_symcache_new(child);
+    REQUIRE(resolver);
+
+    REQUIRE(bcc_symcache_resolve(resolver, (unsigned long long)map_addr,
+        &sym) == 0);
+    REQUIRE(sym.module);
+    // child is PID 1 in its namespace
+    REQUIRE(string(sym.module) == perf_map_path(child));
+    REQUIRE(string("dummy_fn") == sym.name);
+
+    unlink(path.c_str());
+  }
+
+
 
   munmap(map_addr, map_sz);
 }
